backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
authlib==1.2.1
httpx==0.25.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
